feat(SharePage):sharepage desktop layout. refactor(router):remove route to share page, replace by controling with useState
