+++ src/components/pages/TrialComplete/components/OthersTrailInfo.tsx
