+++ src/components/pages/TrialComplete/components/Participants.tsx
@@ -3,14 +3,12 @@ import PlayerCard from "./PlayerCard";
 export default function Participants() {
   return (
     <div className="flex flex-col gap-4 items-center w-2/5 min-w-100 translate-x-4 max-xl:w-full max-xl:translate-x-0 max-xl:px-4">
-      <PlayerCard />  
-      <PlayerCard />  
-      <PlayerCard />  
-      <PlayerCard />  
-      <PlayerCard />  
-      <PlayerCard />  
-
-
+      <PlayerCard />
+      <PlayerCard />
+      <PlayerCard />
+      <PlayerCard />
+      <PlayerCard />
+      <PlayerCard />
     </div>
   );
 }
