+++ src/components/pages/TrialComplete/components/TrialBriefInfo.tsx
