+++ src/routers/routes.tsx
@@ -28,8 +28,6 @@ import CreateTrialForm from "@/components/pages/CreateTrial/components/CreateTri
 
 // Trial Complete
 import TrialComplete from "@/components/pages/TrialComplete";
-import ClosingTrial from "@/components/pages/TrialComplete/components/ClosingTrial";
-import SharePage from "@/components/pages/TrialComplete/components/SharePage";
 
 // Social Pages
 import SocialPages from "@/components/pages/SocialPages";
@@ -119,16 +117,8 @@ export const routes: RouteObject[] = [
       {
         path: "trial-complete",
         element: <TrialComplete />,
-        children: [
-          {
-            index: true,
-            element: <ClosingTrial />,
-          },
-          {
-            path: "share-page",
-            element: <SharePage />,
-          },
-        ],
+
+      
       },
 
       // Social Pages routes
